+++ plugins/Scholarship/src/Model/Table/ApplicationsTable.php
@@ -409,6 +409,7 @@ class ApplicationsTable extends ControllerActionTable
                         'description',
                         'maximum_award_amount',
                         'total_amount',
+                        'duration',
                         'bond',
                         'requirements',
                         'instructions',
@@ -473,6 +474,15 @@ class ApplicationsTable extends ControllerActionTable
     }
 
     // index fields
+    public function onGetFieldLabel(Event $event, $module, $field, $language, $autoHumanize=true)
+    {
+        if ($field == 'scholarship_id') {
+            return __('Scholarship Name');
+        } else {
+            return parent::onGetFieldLabel($event, $module, $field, $language, $autoHumanize);
+        }
+    }
+    
     public function onGetOpenemisNo(Event $event, Entity $entity)
     {
         return $entity->applicant->openemis_no;
@@ -519,6 +529,11 @@ class ApplicationsTable extends ControllerActionTable
     public function onGetMaximumAwardAmount(Event $event, Entity $entity)
     {
         return $entity->scholarship->maximum_award_amount;
+    }
+
+    public function onGetDuration(Event $event, Entity $entity)
+    {
+        return $entity->scholarship->duration . ' ' . __('Years');
     }
 
     public function onGetBond(Event $event, Entity $entity)
@@ -648,6 +663,23 @@ class ApplicationsTable extends ControllerActionTable
         return $attr;
     }
 
+    public function onUpdateFieldDuration(Event $event, array $attr, $action, $request)
+    {
+        if ($action == 'add' || $action == 'edit') {
+            $entity = $attr['entity'];
+
+            $value = '';
+            if (isset($entity->scholarship->duration) && strlen($entity->scholarship->duration) > 0) {
+                $value = $entity->scholarship->duration . ' ' . __('Years');
+            }
+
+            $attr['value'] = $value;
+            $attr['attr']['value'] = $value;
+        }
+
+        return $attr;
+    }    
+
     public function onUpdateFieldInterestRateType(Event $event, array $attr, $action, $request)
     {
         if ($action == 'add' || $action == 'edit') {
@@ -754,8 +786,12 @@ class ApplicationsTable extends ControllerActionTable
             'fieldName' => 'scholarship.maximum_award_amount',
             'attr' => [
                 'require' => false,
-                'label' => $this->addCurrencySuffix('Maximum Award Amount')
+                'label' => $this->addCurrencySuffix('Annual Award Amount')
             ]
+        ]);
+        $this->field('duration', [
+            'type' => 'disabled',
+            'entity' => $entity
         ]);
         $this->field('bond', [
             'type' => 'disabled',