+++ src/Controller/CalendarsController.php
@@ -9,7 +9,6 @@ use Cake\ORM\Entity;
 use Cake\ORM\TableRegistry;
 use Cake\Event\Event;
 use Page\Model\Entity\PageElement;
-
 use App\Controller\PageController;
 
 class CalendarsController extends PageController
@@ -199,7 +198,7 @@ class CalendarsController extends PageController
             // set default academic period to current year
             $academicPeriodId = !is_null($page->getQueryString('academic_period_id')) ? $page->getQueryString('academic_period_id') : $this->AcademicPeriods->getCurrent();
             $page->get('academic_period_id')->setValue($academicPeriodId);
-        } else if ($this->request->is(['post', 'put'])) {
+        } elseif ($this->request->is(['post', 'put'])) {
             $entity = $page->getData();
             $error = $entity->errors();
 
@@ -238,7 +237,7 @@ class CalendarsController extends PageController
 
             $entity->start_date = $startDate;
             $entity->end_date = $endDate;
-        } else if ($this->request->is(['post', 'put'])) {
+        } elseif ($this->request->is(['post', 'put'])) {
             $error = $entity->errors();
 
             if (empty($error)) {