+++ src/Model/Traits/MessagesTrait.php
@@ -517,6 +517,7 @@ trait MessagesTrait
             'current_period' => 'Current Academic Period',
             'next_period' => 'Next Academic Period',
             'success' => 'Students have been promoted',
+            'saveDraftSuccess' => 'Student Promotion successfully saved as draft',
             'noNextGrade' => 'Next grade in the Education Structure is not available in this Institution',
             'reconfirm' => 'Please review the information before proceeding with the operation',
             'noStudentSelected' => 'There are no students selected',
@@ -658,6 +659,15 @@ trait MessagesTrait
             'noStudentSelected' => 'There are no students selected',
             'noLinkedExamCentres' => 'Please contact your administrator to set up available Examination Centres for the selected Examination'
         ],
+        'AssessmentGradingTypes' => [
+            'noGradingOptions' => 'There are no grading options for this grading type'
+        ],
+        'OutcomeGradingTypes' => [
+            'noGradingOptions' => 'There are no grading options for this grading type'
+        ],
+        'CompetencyGradingTypes' => [
+            'noGradingOptions' => 'There are no grading options for this grading type'
+        ],
         'Examinations' => [
             'noExaminationItems' => 'There are no examination items for this examination'
         ],
@@ -743,7 +753,10 @@ trait MessagesTrait
             'publish' => 'The Report Card has been successfully published',
             'publishAll' => 'All generated Report Cards have been published successfully',
             'unpublish' => 'The Report Card has been successfully unpublished',
-            'unpublishAll' => 'All published Report Cards have been unpublished successfully'
+            'unpublishAll' => 'All published Report Cards have been unpublished successfully',
+            'email' => 'The Report Card will be sent in the background',
+            'emailAll' => 'All Report Cards will be sent in the background',
+            'emailInProgress' => 'There is already a email process sending in the background',
         ],
         'RecipientPaymentStructures' => [
             'noApprovedAmount' => 'Please set up Approved Amount for the scholarship'
@@ -835,7 +848,10 @@ trait MessagesTrait
                 ],
                 'institution_provider_id' => [
                     'ruleLinkedSector' => 'Please select a provider linked to the chosen Sector'
-                ]
+                ],
+                'classification' => [
+                    'validClassification' => 'Please enter a valid Classification',
+                ],
             ],
             'InstitutionContacts' => [
                 'email' => [
@@ -999,6 +1015,11 @@ trait MessagesTrait
                 'date_completed' => [
                     'compareWithDateStarted' => 'Date completed should not be earlier than date started'
                 ],
+            ],
+            'FeederOutgoingInstitutions' => [
+                'institution_id' => [
+                    'ruleUnique' => 'Recipient institution must be unique for the same academic period and education grade'
+                ]
             ],
             'Students' => [
                 'student_name' => [
@@ -2041,12 +2062,12 @@ trait MessagesTrait
             ],
             'Applications' => [
                 'requested_amount' => [
-                    'ruleCheckRequestedAmount' => 'Requested amount must not exceed the Maximum Award Amount',
+                    'ruleCheckRequestedAmount' => 'Requested amount must not exceed the Annual Award Amount',
                 ]
             ],
             'ScholarshipRecipients' => [
                 'approved_amount' => [
-                    'comparison' => 'Approved Amount cannot be more than Maximum Award Amount',
+                    'comparison' => 'Approved Award Amount cannot be more than Total Award Amount',
                     'validateDecimal' => 'Value cannot be more than two decimal places',
                     'ruleCheckApprovedWithEstimated' => 'Approved Amount cannot be less than the Estimated Amounts',
                     'ruleCheckApprovedWithDisbursed' => 'Approved Amount cannot be less than the Disbursed Amounts',